requirements.txt
aioredis==2.0.1
anyio==3.7.0
async-timeout==4.0.2
certifi==2023.5.7
charset-normalizer==3.1.0
contourpy==1.1.0
cycler==0.11.0
exceptiongroup==1.1.1
fonttools==4.40.0
greenlet==2.0.1
h11==0.14.0
httpcore==0.17.2
httpx==0.24.0
idna==3.4
importlib-resources==5.12.0
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.25.0
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
playwright==1.33.0
pyee==9.0.4
pyparsing==3.1.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
sniffio==1.3.0
tenacity==8.2.2
tornado==6.3.2
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
zipp==3.15.0
